refactor(frontend): extract CategoriesPage to src/pages/ (Phase 3)
Part of the Phase 3 migration. Route shrinks from 198 to 6 lines.

Hoists the inline DeleteCategoryModal out of the route file to
src/components/categories/DeleteCategoryModal.tsx with a
TODO(migration Phase 5) for the renderDrawer conversion. Hoists the
flattenCategories helper to src/utils/flattenCategories.ts so it can
be reused by other consumers (Select dropdowns elsewhere).

The unused src/components/categories/DeleteCategoryDialog.tsx remains
untouched; Phase 5 will reconcile the two.
